requirements.txt
discord.py
requests
feedparser